requirements.txt
phidata
fastapi
tavily-python
crawl4ai